refactor(lists): extract the relay edge out of the curated list god file
#6932 froze lib/services/curated_list_service.dart at 1848 lines while
this branch was in review, and the private-lists work had taken it to
2289. The ratchet is explicit that a baseline bump cannot approve growth,
so this moves work out of the file rather than raising the ceiling.

CuratedListRelayGateway takes everything the service did at the relay
edge that does not touch its local list state: public-list reads, the
event-to-list codec, NIP-44 sealing and unsealing, and the NIP-09
deletion and redaction publishes. The service keeps the cache and calls
through the gateway; its public read methods stay as thin delegates so no
call site moves. Behaviour is unchanged -- the 241 existing curated-list
service tests pass untouched -- and the gateway gets its own tests for
the sealing guards, unseal classification, and redaction tag shape.

The service lands at 1757 lines, under the frozen ceiling. The full move
into curated_list_repository remains the #4338 end state; this is ratchet
compliance, not that migration.
